+++ crates/backend/src/typegen/struct.rs
@@ -2,11 +2,11 @@ use std::collections::HashMap;
 use std::vec::Vec;
 use std::{cell::RefCell, iter};
 
-use super::{add_alias, ToTypeDef, TypeDef};
+use convert_case::Casing;
+
+use super::{add_alias, format_js_property_name, ty_to_ts_type, ToTypeDef, TypeDef};
 use crate::typegen::JSDoc;
-use crate::{
-  format_js_property_name, ty_to_ts_type, NapiImpl, NapiStruct, NapiStructField, NapiStructKind,
-};
+use crate::{NapiImpl, NapiStruct, NapiStructField, NapiStructKind};
 
 thread_local! {
   pub(crate) static TASK_STRUCTS: RefCell<HashMap<String, String>> = Default::default();
@@ -198,7 +198,12 @@ impl NapiStruct {
         .map(|variant| {
           let def = iter::once(format!(
             "{}: '{}'",
-            structured_enum.discriminant, variant.name
+            structured_enum.discriminant,
+            if let Some(case) = structured_enum.discriminant_case {
+              variant.name.to_string().to_case(case)
+            } else {
+              variant.name.to_string()
+            }
           ))
           .chain(
             variant