+++ examples/napi/__tests__/values.spec.ts
@@ -227,6 +227,8 @@ import {
   callFinallyOnPromise,
   StructuredKind,
   validateStructuredEnum,
+  StructuredKindLowercase,
+  validateStructuredEnumLowercase,
   createArraybuffer,
   getBufferSlice,
   createExternalBufferSlice,
@@ -287,6 +289,7 @@ import {
   indexSetToRust,
   indexSetToJs,
   intoUtf8,
+  withAbortSignalHandle,
 } from '../index.cjs'
 // import other stuff in `#[napi(module_exports)]`
 import nativeAddon from '../index.cjs'
@@ -445,6 +448,30 @@ test('structured enum', (t) => {
   t.deepEqual(tuple, validateStructuredEnum(tuple))
   t.throws(() => validateStructuredEnum({ type2: 'unknown' } as any))
   t.throws(() => validateStructuredEnum({ type2: 'Greeting' } as any))
+
+  const hello2: StructuredKindLowercase = {
+    type: 'hello',
+  }
+  const greeting2: StructuredKindLowercase = {
+    type: 'greeting',
+    name: 'Napi-rs',
+  }
+  const birthday2: StructuredKindLowercase = {
+    type: 'birthday',
+    name: 'Napi-rs',
+    age: 10,
+  }
+  const tuple2: StructuredKindLowercase = {
+    type: 'tuple',
+    field0: 1,
+    field1: 2,
+  }
+  t.deepEqual(hello2, validateStructuredEnumLowercase(hello2))
+  t.deepEqual(greeting2, validateStructuredEnumLowercase(greeting2))
+  t.deepEqual(birthday2, validateStructuredEnumLowercase(birthday2))
+  t.deepEqual(tuple2, validateStructuredEnumLowercase(tuple2))
+  t.throws(() => validateStructuredEnumLowercase({ type: 'unknown' } as any))
+  t.throws(() => validateStructuredEnumLowercase({ type: 'greeting' } as any))
 })
 
 test('function call', async (t) => {
@@ -1458,6 +1485,20 @@ test('async task with different resolved values', async (t) => {
   }
   const r2 = await asyncResolveArray(2)
   t.deepEqual(r2, [0, 1])
+})
+
+AbortSignalTest('with abort signal handle', async (t) => {
+  const ctrl = new AbortController()
+  const promise = withAbortSignalHandle(ctrl.signal)
+  try {
+    ctrl.abort()
+    const ret = await promise
+    t.is(ret, 999)
+  } catch (err: unknown) {
+    // sometimes on CI, the scheduled task is able to abort
+    // so we only allow it to throw AbortError
+    t.is((err as Error).message, 'AbortError')
+  }
 })
 
 AbortSignalTest('abort resolved task', async (t) => {